get one do cartao finalizado

+++ users-vem/src/main/java/com/urbanape/usersvem/repository/CartaoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.urbanape.usersvem.model.CartaoModel;
 
-public interface CartaoRepository extends JpaRepository<CartaoModel, Long>{
+public interface CartaoRepository extends JpaRepository<CartaoModel, Integer>{
 }
